Make pictures render
