+++ js/main.js
@@ -1,57 +1,8 @@
 /* eslint-disable no-unused-vars */
-import {calculateRandomNumber, getRandomElement} from './util.js';
-import {NAMES, COMMENTS, DESCRIPTION} from './data.js';
-
-// Задаем количество фотографий и диапазон лайков
-const quantityPictures = 25;
-const likes = {
-  min: 15,
-  max: 200,
-};
+import './util.js';
+import './data.js';
+import {resultRender} from './rendering-pictures.js';
 
 
-/**
- * Функция, создающая комментарий к фотографии.
- *
- * @param {number} id - случайное число. Идентификаторы не должны повторяться.
- * @returns {object} - комментарий в виде объекта.
- */
-function createComment(id){
-  return {
-    id: calculateRandomNumber(1, 500),
-    avatar: `img/avatar-${id}.svg`,
-    message: getRandomElement(COMMENTS),
-    name: getRandomElement(NAMES)
-  };
-}
-
-
-/**
- * Функция, создающая комментарий к фотографии.
- *
- * @param {number} len - длинна создаваемого массива(пустого).
- * @param {function} func - функция, которая будет заполнять массив.
- * @returns {array}
- */
-const createArrayObj = (len, func) => Array.from({length: len}).map((value, index) => func(index + 1));
-
-
-// Массив из quantityPictures сгенерированных объектов.
-const resultArray = createArrayObj(quantityPictures, createDescription);
-
-
-/**
- * Функция, создающая комментарий к фотографии.
- *
- * @param {number} id - случайное число. Идентификаторы не должны повторяться.
- * @returns {object} - информация о фотографии.
- */
-function createDescription(id){
-  return {
-    id: id,
-    url: `photos/${id}.jpg`,
-    description: getRandomElement(DESCRIPTION),
-    likes: calculateRandomNumber(likes.min, likes.max),
-    comments: createArrayObj(getRandomElement(1, 6), createComment),
-  };
-}
+// eslint-disable-next-line no-console
+console.log(resultRender);
